refactor(wire): extract document adapter lifecycle
## Summary

- Move the document adapter and its two preflight validators into the
  771-line `_wire_document.py` lifecycle owner.
- Reduce `_wire.py` to 1,531 lines and retarget Vulture owner paths.

## Boundaries and compatibility

- Preserve all 11 function bodies and the private body slot exactly.
- Preserve generated names, modules, closures, and lifecycle behavior.
- Keep the duplicate schema-hook preflight until the cleanup issue.
- Leave every automatic exclusion path untouched.

## Investigation

- Source and AST comparisons found no missing or changed definitions.
- The document module depends only on contract and compiler leaves.
- Three Vulture expectations now name the semantic document owner.

## Validation

- `uv run make ci`: passed; 859 tests; 95.88% coverage.
- Python 3.12 with Pydantic 2.12.3: 10 focused tests passed.
- Ruff, ty, strict mypy, Vulture, docs, and package build passed.
- Immutable v0.3.0 compatibility contract passed.

Closes #121
Refs #95
